# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse:
    properties:
      error:
        example: unauthorized
        type: string
      request_id:
        example: abc123
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.ListResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.Meta'
      results: {}
      status:
        enum:
        - success
        example: success
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.Meta:
    properties:
      pagination:
        $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.Pagination'
      version:
        example: "1"
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse:
    properties:
      meta:
        $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.Meta'
      result: {}
      status:
        enum:
        - success
        example: success
        type: string
    type: object
  github_com_Yab1_golang-template_internal_platform_httpx.Pagination:
    properties:
      limit:
        example: 20
        type: integer
      next_cursor:
        example: eyJ0IjoiMjAyNi0wOS0xOFQxMjowMDowMFoiLCJpIjoiLi4uIn0
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 42
        type: integer
    type: object
...
  modules_post.CreatePostPayload:
    properties:
      content:
        maxLength: 10000
        type: string
      tags:
        items:
          type: string
        maxItems: 5
        type: array
      title:
        maxLength: 255
        type: string
    required:
    - content
    - title
    type: object
  modules_post.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      reference_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      user_id:
        type: string
      version:
        type: integer
    type: object
  modules_post.UpdatePostPayload:
    properties:
      content:
        maxLength: 10000
        type: string
      tags:
        items:
          type: string
        maxItems: 5
        type: array
      title:
        maxLength: 255
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: yeabsera.dev@gmail.com
    name: Yeabsera
    url: https://yeabsera-dev.vercel.app/
  description: 'REST API template: JWT access+refresh, RBAC, posts CRUD, files (local/minio/s3),
    mail, Redis rate limiting, human reference ids.'
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.google.com/policies/terms/
  title: Golang Template API
  version: 0.0.1
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: List posts with pagination, sorting, search, and tag filters (public)
      parameters:
      - default: 20
        description: Page size (1-100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset (ignored when cursor is set)
        in: query
        name: offset
        type: integer
      - description: Opaque keyset cursor (created_at+id). Requires sort_by=created_at
        in: query
        name: cursor
        type: string
      - default: created_at
        description: Sort field
        enum:
        - created_at
        - updated_at
        - title
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - description: Search title/content
        in: query
        name: search
        type: string
      - description: Comma-separated tags (AND / contains all)
        in: query
        name: tags
        type: string
      - description: Filter by author UUID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ListResponse'
            - properties:
                results:
                  items:
                    $ref: '#/definitions/modules_post.Post'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a post as the authenticated user
      parameters:
      - description: Post payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/modules_post.CreatePostPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse'
            - properties:
                result:
                  $ref: '#/definitions/modules_post.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a post
      tags:
      - posts
  /posts/{postID}:
    delete:
      consumes:
      - application/json
      description: Soft-delete (sets deleted_at). Owner or admin+.
      parameters:
      - description: Post UUID or reference_id (e.g. GTL-PST-A7K2M)
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get a post by ID (public)
      parameters:
      - description: Post UUID or reference_id (e.g. GTL-PST-A7K2M)
        in: path
        name: postID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse'
            - properties:
                result:
                  $ref: '#/definitions/modules_post.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      summary: Get a post
      tags:
      - posts
    patch:
      consumes:
      - application/json
      description: Owner or moderator+ can update. Optimistic locking via version.
      parameters:
      - description: Post UUID or reference_id (e.g. GTL-PST-A7K2M)
        in: path
        name: postID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/modules_post.UpdatePostPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ObjectResponse'
            - properties:
                result:
                  $ref: '#/definitions/modules_post.Post'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_Yab1_golang-template_internal_platform_httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access JWT (or AUTH_DEV_TOKEN
      in development).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"